% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.adj.R
\name{adjust_linear}
\alias{adjust_linear}
\title{Linearly adjust data based on two data points of know quality.}
\usage{
adjust_linear(raw, calib, rowfirst, rowlast, rownums = 1:length(raw))
}
\arguments{
\item{raw}{(numeric) Numeric vector of raw data, sorted by date time ascending.}

\item{calib}{(numeric) Numeric vector that is blank except for the calibration values.}

\item{rowfirst}{(numeric) Rownumber of the first calibration value.}

\item{rowlast}{(numeric) Rownumber of the second (and last) calibration value.}

\item{rownums}{(numeric) Vector of row numbers. Defaults to 1:length(raw)}
}
\value{
A vector with caliberated values. Adjustment is linear.
}
\description{
Linearly adjust data based on two data points of know quality.
}
