% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc.R
\name{alloc_mat_mat_mul}
\alias{alloc_mat_mat_mul}
\title{Allocate empty matrix for the result of the matrix multiplication \code{A * B}}
\usage{
alloc_mat_mat_mul(A, B, ta = FALSE, tb = FALSE)
}
\arguments{
\item{A, B}{Primary multiplication matrices}

\item{ta, tb}{Should matrix be transposed? Default: FALSE}
}
\value{
Uninitialized matrix of the correct size to hold the result
}
\description{
Allocate empty matrix for the result of the matrix multiplication \code{A * B}
}
\examples{
A <- matrix(1, 2, 4)
B <- matrix(1, 4, 7)
C <- alloc_mat_mat_mul(A, B)
is.matrix(C)
dim(C)
}
