% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc.R
\name{alloc_mat_vec_mul}
\alias{alloc_mat_vec_mul}
\title{Allocate empty vector for the result of the matrix-vector multiplication \code{A * x}}
\usage{
alloc_mat_vec_mul(A, x)
}
\arguments{
\item{A}{Matrix}

\item{x}{vector}
}
\value{
Uninitialized vector of the correct size to hold the result of \code{A * x}
}
\description{
Allocate empty vector for the result of the matrix-vector multiplication \code{A * x}
}
\examples{
A <- matrix(1, 2, 4)
x <- matrix(1, 4, 1)
y <- alloc_mat_vec_mul(A, x)
dim(y)
br_mat_vec_mul(y, A, x)
y
# Compare to base R
A \%*\% x
}
