% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc.R
\name{alloc_matrix}
\alias{alloc_matrix}
\title{Allocate empty matrix of the given dimensions.}
\usage{
alloc_matrix(nrow, ncol)
}
\arguments{
\item{nrow, ncol}{dimensions}
}
\value{
Matrix of the requested dimensions. Values are \emph{not} initialised.
}
\description{
The values are in the matrix are \emph{not} initialised to zero.
}
\examples{
m <- alloc_matrix(nrow = 3, ncol = 2)
is.matrix(m)
dim(m)
}
