% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-matrix.R
\name{alloc_matrix_mul}
\alias{alloc_matrix_mul}
\title{Allocate empty matrix for the result of the matrix multiplication \code{A * B}}
\usage{
alloc_matrix_mul(A, B)
}
\arguments{
\item{A, B}{matrices to be multiplied}
}
\value{
Uninitialized matrix of the correct size to hold the result
}
\description{
Allocate empty matrix for the result of the matrix multiplication \code{A * B}
}
\examples{
A <- matrix(1, 2, 4)
B <- matrix(1, 4, 7)
C <- alloc_matrix_mul(A, B)
is.matrix(C)
dim(C)
}
