% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-ops-unary.R
\name{br_abs}
\alias{br_abs}
\alias{br_sqrt}
\alias{br_floor}
\alias{br_ceil}
\alias{br_trunc}
\alias{br_round}
\alias{br_exp}
\alias{br_log}
\alias{br_cos}
\alias{br_sin}
\alias{br_tan}
\alias{br_not}
\alias{br_expm1}
\alias{br_log1p}
\alias{br_acos}
\alias{br_asin}
\alias{br_atan}
\alias{br_acosh}
\alias{br_asinh}
\alias{br_atanh}
\alias{br_cosh}
\alias{br_sinh}
\alias{br_tanh}
\alias{br_cospi}
\alias{br_sinpi}
\alias{br_tanpi}
\alias{br_sign}
\alias{br_cumsum}
\alias{br_cumprod}
\alias{br_cummax}
\alias{br_cummin}
\alias{br_log2}
\alias{br_log10}
\alias{br_is_na}
\title{Math ops}
\usage{
br_abs(x)

br_sqrt(x)

br_floor(x)

br_ceil(x)

br_trunc(x)

br_round(x, digits)

br_exp(x)

br_log(x)

br_cos(x)

br_sin(x)

br_tan(x)

br_not(x)

br_expm1(x)

br_log1p(x)

br_acos(x)

br_asin(x)

br_atan(x)

br_acosh(x)

br_asinh(x)

br_atanh(x)

br_cosh(x)

br_sinh(x)

br_tanh(x)

br_cospi(x)

br_sinpi(x)

br_tanpi(x)

br_sign(x)

br_cumsum(x)

br_cumprod(x)

br_cummax(x)

br_cummin(x)

br_log2(x)

br_log10(x)

br_is_na(x)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}

\item{digits}{number of decimal places}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
Math ops
}
\examples{
x <- c(-1, 0, 1)
br_abs(x)
x

}
