% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-ops-binary.R
\name{br_add}
\alias{br_add}
\alias{br_sub}
\alias{br_mul}
\alias{br_div}
\alias{br_pow}
\alias{br_eq}
\alias{br_ne}
\alias{br_lt}
\alias{br_le}
\alias{br_gt}
\alias{br_ge}
\alias{br_and}
\alias{br_or}
\alias{br_rem}
\alias{br_idiv}
\alias{br_max}
\alias{br_min}
\title{In-situ ops}
\usage{
br_add(x, y)

br_sub(x, y)

br_mul(x, y)

br_div(x, y)

br_pow(x, y)

br_eq(x, y)

br_ne(x, y)

br_lt(x, y)

br_le(x, y)

br_gt(x, y)

br_ge(x, y)

br_and(x, y)

br_or(x, y)

br_rem(x, y)

br_idiv(x, y)

br_max(x, y)

br_min(x, y)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}

\item{y}{Either scalar numeric values or numeric vectors of the same 
length as \code{x}.}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
In-situ ops
}
\examples{
# x <- x + y
x <- as.numeric(1:10)
y <- as.numeric(1:10)
br_add(x, y)
x
# x <- x + 1
br_add(x, 1)
x
}
