% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-copy.R
\name{br_copy_if}
\alias{br_copy_if}
\title{Copy all or part of one vector into another}
\usage{
br_copy_if(x, y, where)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}

\item{y}{Either scalar numeric value or numeric vector of the same 
length as \code{x}.}

\item{where}{numeric vector the same length as \code{x} containing logical
values where 0 indicates false, and all non-zero values
indicate true}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
Copy all or part of one vector into another
}
\examples{
x <- c(-1, 0, 1)

# Where is x < 0?
xneg <- duplicate(x)
br_lt(xneg, 0) #  xneg <- (xneg < 0)
xneg

br_copy_if(x, -99, xneg)
x

}
