% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-dist.R
\name{br_dist2}
\alias{br_dist2}
\alias{br_dist3}
\title{Distance calculation}
\usage{
br_dist2(x1, y1, x2, y2)

br_dist3(x1, y1, z1, x2, y2, z2)
}
\arguments{
\item{x1, y1, z1, x2, y2, z2}{numeric vectors all the same length}
}
\value{
\code{x1} modified in-place and returned invisibly
}
\description{
Distance calculation
}
\examples{
N <- 10
x1 <- rep(1, N)
y1 <- runif(N)
z1 <- runif(N)
x2 <- runif(N)
y2 <- runif(N)
z2 <- runif(N)
br_dist2(x1, y1, x2, y2)
x1
}
