% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-fma.R
\name{br_fmadd}
\alias{br_fmadd}
\alias{br_fmsub}
\alias{br_fnmadd}
\alias{br_fnmsub}
\title{Fused multiply add}
\usage{
br_fmadd(x, a, b)

br_fmsub(x, a, b)

br_fnmadd(x, a, b)

br_fnmsub(x, a, b)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}

\item{a, b}{Either scalar numeric values or numeric vectors of the same 
length as \code{x}.}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
Performs a compound calculation equivalent to \code{x <- x * a + b} using
the best precision from the math library.
}
\details{
\describe{
  \item{\code{br_fmadd()}}{Equivalent to: \code{x <- x * a + b}}
  \item{\code{br_fmsub()}}{Equivalent to: \code{x <- x * a - b}}
  \item{\code{br_fnmadd()}}{Equivalent to: \code{x <- -x * a + b}}
  \item{\code{br_fnmsub()}}{Equivalent to: \code{x <- -x * a - b}}
}
}
\examples{
x <- c(1, 2, 3)
a <- c(2, 3, 4)
b <- c(10, 10, 10)
br_fmadd(x, a, b)
x
}
