% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-dist.R
\name{br_hypot}
\alias{br_hypot}
\alias{br_hypot2}
\alias{br_hypot3}
\title{Hypotenuse length calculation (distance from origin)}
\usage{
br_hypot(x, y)

br_hypot2(x, y)

br_hypot3(x, y, z)
}
\arguments{
\item{x, y, z}{numeric vectors all the same length}
}
\value{
\code{x1} modified in-place and returned invisibly
}
\description{
Hypotenuse length calculation (distance from origin)
}
\examples{
# Distance from origin in 3D
N <- 10
x <- rep(1, N)
y <- runif(N)
br_hypot2(x, y)
x

# Distance from origin in 2D
x <- rep(1, N)
z <- runif(N)
br_hypot3(x, y, z)
}
