% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-matrix.R
\name{br_mat_col_get}
\alias{br_mat_col_get}
\alias{br_mat_col_set}
\alias{br_mat_row_get}
\alias{br_mat_row_set}
\title{Get/set a column or row of a matrix}
\usage{
br_mat_col_get(mat, i, vec)

br_mat_col_set(mat, i, vec)

br_mat_row_get(mat, i, vec)

br_mat_row_set(mat, i, vec)
}
\arguments{
\item{mat}{matrix}

\item{i}{index of row or column}

\item{vec}{vector space. For \emph{get} operations, vector length must 
match the row or column length of the matirx.  When doing a \code{set}
operation, \code{vec} may also be a single value (which will be repeated
to fill the row/column)}
}
\value{
None. For \code{get} operations, \code{vec} is modified by-reference 
     and returned invisibly.  For \code{set} operations, \code{mat} is modified
     by reference and returned invisibly.
}
\description{
Get/set a column or row of a matrix
}
\examples{
m <- matrix(as.numeric(1:6), 2, 3)
m
v <- alloc_n(nrow(m))
br_mat_col_get(m, 2, v)
v
}
