% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-dist.R
\name{br_mat_dist2}
\alias{br_mat_dist2}
\alias{br_mat_dist3}
\title{Distance calculation}
\usage{
br_mat_dist2(d, mat1, mat2)

br_mat_dist3(d, mat1, mat2)
}
\arguments{
\item{d}{pre-allocated vector to hold the result}

\item{mat1, mat2}{matrices holding (x, y) or (x, y, z) coordinates}
}
\value{
None. Argument \code{d} modified in-place and returned invisibly
}
\description{
Distance calculation
}
\examples{
N <- 10
x1 <- rep(1, N)
y1 <- runif(N)
z1 <- runif(N)
x2 <- runif(N)
y2 <- runif(N)
z2 <- runif(N)
mat1 <- cbind(x1, y1, z1)
mat2 <- cbind(x2, y2, z2)
d <- alloc_n(N)
br_mat_dist2(d, mat1, mat2)
d
}
