% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-dist.R
\name{br_mat_hypot2}
\alias{br_mat_hypot2}
\alias{br_mat_hypot3}
\title{Hypotenuse length calculation (distance from origin)}
\usage{
br_mat_hypot2(d, mat)

br_mat_hypot3(d, mat)
}
\arguments{
\item{d}{pre-allocated numeric vector with length matching \code{nrow(mat)}}

\item{mat}{numeric matrix}
}
\value{
None. Argument \code{d} modified in-place and returned invisibly.
}
\description{
Hypotenuse length calculation (distance from origin)
}
\examples{
# Distance from origin in 3D
N <- 10
x <- rep(1, N)
y <- runif(N)
z <- runif(N)
mat <- cbind(x, y, z)
d1 <- alloc_n(N)
br_mat_hypot3(d1, mat)
d1

# Compare to base R
d2 <- sqrt(x * x + y * y + z * z)
all.equal(d1, d2)
}
