% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-matrix.R
\name{br_mat_mat_mul}
\alias{br_mat_mat_mul}
\title{Matrix-matrix multiplication}
\usage{
br_mat_mat_mul(C, A, B, alpha = 1, beta = 0, ta = FALSE, tb = FALSE)
}
\arguments{
\item{C}{Matrix which will be modified by-reference}

\item{A, B}{Primary multiplication matrices}

\item{alpha}{Scaling factor for \code{A * B}. Default: 1.0}

\item{beta}{Scaling factor for \code{C}. Default: 0.0}

\item{ta, tb}{Should matrix be transposed? Default: FALSE}
}
\value{
C is modified by-reference and returned invisibly
}
\description{
This function exposes the general matrix multiplication operation from R's
included BLAS.  \code{C = alpha * A * B + beta * C}
}
\examples{
A <- matrix(as.numeric(1:32), 8, 4)
B <- matrix(as.numeric(1:24), 4, 6)  
C <- alloc_mat_mat_mul(A, B)

br_mat_mat_mul(C, A, B) # By reference. Overwriting 'C'
C

A \%*\% B  # Compare to base R
}
