% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-matrix.R
\name{br_mat_mat_mul_bsq}
\alias{br_mat_mat_mul_bsq}
\title{Matrix-matrix multiplication when B is a square matrix.}
\usage{
br_mat_mat_mul_bsq(A, B, alpha = 1)
}
\arguments{
\item{A, B}{Matrices}

\item{alpha}{Scaling factor for \code{A * B}. Default: 1.0}
}
\value{
A is modified by-reference and returned invisibly
}
\description{
This function exposes the general matrix multiplication operation from R's
included BLAS.  \code{A <- A * B}
}
\examples{
A <- matrix(as.numeric(1:32), 8, 4)
B <- matrix(as.numeric(1:16), 4, 4)  

A \%*\% B  # Base R result.

br_mat_mat_mul_bsq(A, B) # By reference. Overwriting 'A'
A
}
