% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-dist.R
\name{br_mat_normalise2}
\alias{br_mat_normalise2}
\alias{br_mat_normalise3}
\title{Normalise matrix of (x,y) coordinates to length 1}
\usage{
br_mat_normalise2(mat)

br_mat_normalise3(mat)
}
\arguments{
\item{mat}{numeric matrix}
}
\value{
Input matrix modified in-place and returned invisibly.
}
\description{
These functions will normalise matrices of (x, y) or (x, y, z) coordinates so that
the represented vector quantity have a length of 1.
}
\examples{
N <- 10
x <- runif(N)
y <- runif(N)
mat <- cbind(x, y)
br_mat_normalise2(mat)
# Lengths should now be 1.
sqrt(mat[,1]^2 + mat[,2]^2)
}
