% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-matrix.R
\name{br_mat_roll}
\alias{br_mat_roll}
\title{Roll elements of a matrix}
\usage{
br_mat_roll(mat, rows = 0, cols = 0)
}
\arguments{
\item{mat}{matrix}

\item{rows, cols}{Number of rows and columns to roll}
}
\value{
None. Matrix is modified by-reference and returned invisibly
}
\description{
Roll elements of a matrix
}
\examples{
m <- matrix(as.numeric(1:6), 2, 3)
m
br_mat_roll(m, rows = 3)
m
}
