% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-matrix.R
\name{br_mat_transpose}
\alias{br_mat_transpose}
\title{Transpose matrix}
\usage{
br_mat_transpose(mat)
}
\arguments{
\item{mat}{matrix}
}
\value{
None. Matrix is modified by-reference and returned invisibly
}
\description{
Transpose matrix
}
\examples{
m <- matrix(as.numeric(1:6), 2, 3)
m
br_mat_transpose(m)
m
}
