% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-matrix.R
\name{br_mat_vec_mul}
\alias{br_mat_vec_mul}
\title{Matrix-vector multiplication}
\usage{
br_mat_vec_mul(y, A, x, alpha = 1, beta = 0)
}
\arguments{
\item{y}{Vector which will be modified by-reference}

\item{A}{Matrix}

\item{x}{vector}

\item{alpha}{Scaling factor for \code{A * x}. Default: 1.0}

\item{beta}{Scaling factor for \code{y}. Default: 0.0}
}
\value{
y is modified by-reference and returned invisibly
}
\description{
This function exposes the general matrix-vector multiplication operation from R's
included BLAS.  \code{y = alpha * A * x + beta * y}
}
\examples{
A <- matrix(1, 3, 5)
x <- rep(1, 5)
y <- rep(0, 3)

# Calculate. By-reference. Overwriting 'y'
br_mat_vec_mul(y, A, x)
y

# Compare to R's method
A \%*\% x

}
