% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-matrix.R
\name{br_mat_vec_mul_asq}
\alias{br_mat_vec_mul_asq}
\title{Matrix-vector multiplication when A is a square matrix}
\usage{
br_mat_vec_mul_asq(A, x, alpha = 1)
}
\arguments{
\item{A}{Matrix}

\item{x}{vector. This will be modified by-reference}

\item{alpha}{Scaling factor for \code{A * x}. Default: 1.0}
}
\value{
x is modified by-reference and returned invisibly
}
\description{
This function calculates matrix-vector multiplication operation when 'A'
is a square matrix: \code{x <- alpha * A * x}
}
\examples{
A <- matrix(1, 4, 4)
x <- rep(1, 4)

# Compare to R's method
A \%*\% x

# Calculate. By-reference. Overwriting 'y'
br_mat_vec_mul_asq(A, x)
x

}
