% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-matrix.R
\name{br_mul_mat_mat}
\alias{br_mul_mat_mat}
\title{Matrix-matrix multiply}
\usage{
br_mul_mat_mat(A, B, alpha = 1, A_transpose = FALSE, B_transpose = FALSE)
}
\arguments{
\item{A, B}{Matrices}

\item{alpha}{Scaling factor for \code{A * B}. Default: 1.0}

\item{A_transpose, B_transpose}{Should matrix be transposed? Default: FALSE}
}
\value{
A is modified by-reference and returned invisibly
}
\description{
This function exposes the general matrix multiplication operation from R's
included BLAS.  \code{A <- A * B}
}
\examples{
A <- matrix(1, 8, 4)
B <- matrix(2, 4, 4)  

A \%*\% B  # Base R result.

br_mul_mat_mat(A, B) # By reference. Overwriting 'A'
A
}
