% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-dist.R
\name{br_normalise2}
\alias{br_normalise2}
\alias{br_normalise3}
\title{Normalise vectors to length 1}
\usage{
br_normalise2(x, y)

br_normalise3(x, y, z)
}
\arguments{
\item{x, y, z}{numeric vectors all the same length}
}
\value{
None. \code{x, y and z} modified in-place.
}
\description{
These functions will normalise the (x, y) or (x, y, z) coordinates so that
the represented vector quantity has a length of 1.   All arguments will 
be modified in place.
}
\examples{
N <- 10
x <- runif(N)
y <- runif(N)
br_normalise2(x, y)
x
y
sqrt(x^2 + y^2)
}
