% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-copy.R
\name{br_replace}
\alias{br_replace}
\title{Replace elements of one vector with another}
\usage{
br_replace(x, xi, y, yi = NULL)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}

\item{xi}{destination indices in \code{x} for the replacement operation}

\item{y}{A numeric vector}

\item{yi}{source indices in \code{y} for the replacement. If NULL (the default),
then values are taken sequentially from \code{y}}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
This is similar to \code{br_copy()} except that \code{xi} and \code{yi}
are vectors of indices into \code{x} and \code{y}, respectively.
}
\examples{
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# sequential y indices assumed if 'yi' is NULL (the default)
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
x <- c(1, 2, 3, 4, 5)
y <- c(99, 98)
xi <- c(1L, 3L)

br_replace(x, xi, y)
x


#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
# explicit y
#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
x <- c(1, 2, 3, 4, 5)
y <- c(99, 98, 97, 96)
xi <- c(1L, 3L)
yi <- c(4L, 3L)

br_replace(x, xi, y, yi)
x
}
