% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-rev.R
\name{br_rev}
\alias{br_rev}
\title{Reverse a vector in place}
\usage{
br_rev(x)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
Reverse a vector in place
}
\examples{
x <- as.numeric(1:10)
br_rev(x)
x
}
