% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-roll.R
\name{br_roll}
\alias{br_roll}
\title{Roll a vector}
\usage{
br_roll(x, dist)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}

\item{dist}{Distance to roll. Can be postive or negative}
}
\value{
None. \code{x} modified in-place and returned invisibly
}
\description{
Roll a vector
}
\examples{
x <- c(1, 2, 3, 4, 5)
br_roll(x, 2)
x
br_roll(x, -2)
x 
br_roll(x, -3)
}
