% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-runif.R
\name{br_runif}
\alias{br_runif}
\title{Fill a vector with uniform random values}
\usage{
br_runif(x, lower = 0, upper = 1)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}

\item{lower, upper}{limits of random numbers generated. Default: 0, 1}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
Random numbers are generated using a Lehmer RNG (also known as Park-Miller RNG).
Note: The standard R RNG is only used to seed the Lehmer RNG - which means
that \code{set.seed()} can be used for generating repeatable streams.
}
\examples{
set_seed_lehmer(1)
x <- alloc_n(10)
br_runif(x)
x

}
