% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-seq.R
\name{br_seq}
\alias{br_seq}
\title{Fill vector with a numeric sequence}
\usage{
br_seq(x, from = 1, to = NULL, step = NULL)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}

\item{from}{starting value in sequence. Default: 1}

\item{to}{ending value in sequence. Default: NULL.  If NULL, then 
\code{to = length(x) + from - 1}}

\item{step}{step size. This value is ignored if \code{to} is set.}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
Fill vector with a numeric sequence
}
\examples{
# fill 1:10
x <- alloc_n(10)
br_seq(x)
x


# fill from 1 to 100
x <- alloc_n(10)
br_seq(x, from = 1, to = 100)
x

# fill from 0 in steps of 2.5
x <- alloc_n(10)
br_seq(x, from = 0, step = 0.25)
x

}
