% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-shuffle.R
\name{br_shuffle}
\alias{br_shuffle}
\title{Shuffle a vector in place}
\usage{
br_shuffle(x)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
Uses fisher-yates.
}
\details{
The \emph{fast} variant of this function uses a xoshiro256++ PRNG which is
much faster than sourcing randomness from R's random number generator.
}
\examples{
set.seed(1)
set_seed_lehmer()
x <- as.numeric(1:10)
br_shuffle(x)
x
}
