% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-sort.R
\name{br_sort}
\alias{br_sort}
\title{Sort a vector of real values in-situ}
\usage{
br_sort(x, decreasing = FALSE)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}

\item{decreasing}{Logical. Sort values in decreasing order? default FALSE}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
Sort a vector of real values in-situ
}
\examples{
set.seed(1)
x <- sample(as.numeric(1:10))
br_sort(x)
x
}
