% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-zzz.R
\docType{data}
\name{by_reference}
\alias{by_reference}
\title{List of all functions}
\format{
An object of class \code{list} of length 66.
}
\usage{
by_reference
}
\description{
List of all functions
}
\examples{
x <- as.numeric(1:10)
with(by_reference, {x + 1; x * 2})
x
}
\keyword{datasets}
