% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_salinity.R
\name{calculate_salinity}
\alias{calculate_salinity}
\title{Convert Conductivity to Salinity}
\usage{
calculate_salinity(cond, temp, ref_cond = 42.9)
}
\arguments{
\item{cond}{(numeric) Numeric vector of conductivity values in miliSeimens/m^2/s.}

\item{temp}{(numeric) Numeric vector of temperature values in degree Celsius.}

\item{ref_cond}{(numeric) Reference conductivity value, defaults to 42.9 for conductivity measured in miliSeimens/m^2/s.}
}
\value{
(numeric) Vector of salinity values calculated from each pair of cond and temp values in input vectors. Unit varies depend on unit for conductivity, defaults to  PSU.
}
\description{
Calculate salinity from conductivitity and temperature, using equations from UNESCO 1983. Taken from James Douglass's Excel sheet.
}
