% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_polytomies.R
\name{check_polytomies}
\alias{check_polytomies}
\title{Detects and handles unresolved nodes that would confound ASR}
\usage{
check_polytomies()
}
\description{
Detects and handles unresolved nodes that would confound ASR
}
