% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_tips.R
\name{classify_tips}
\alias{classify_tips}
\title{Assigns tips as island endemic, non-endemic, or widespread based on locality data}
\usage{
classify_tips()
}
\description{
Assigns tips as island endemic, non-endemic, or widespread based on locality data
}
