% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondtoSal_miliS.R
\name{Cond.to.Sal}
\alias{Cond.to.Sal}
\title{Convert Conductivity to Salinity
using equations from UNESCO 1983. Taken from James Douglass's Excel sheet}
\usage{
Cond.to.Sal(cond, temp)
}
\arguments{
\item{cond}{Conductivity in miliSeimens/m^2/s}

\item{temp}{Temperature in Celsius}
}
\description{
Convert Conductivity to Salinity
using equations from UNESCO 1983. Taken from James Douglass's Excel sheet
}
