% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_radiation_bursts.R
\name{detect_radiation_bursts}
\alias{detect_radiation_bursts}
\title{Identifies nodes where multiple sequential in-situ speciation events occur within a defined time window. Useful for adaptive radiations}
\usage{
detect_radiation_bursts()
}
\description{
Identifies nodes where multiple sequential in-situ speciation events occur within a defined time window. Useful for adaptive radiations
}
