% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversity_through_time.R
\name{diversity_through_time}
\alias{diversity_through_time}
\title{Reconstructs how in-situ vs. colonization diversity accumulated through geological time on each island (need to include island ages somewhere)}
\usage{
diversity_through_time()
}
\description{
Reconstructs how in-situ vs. colonization diversity accumulated through geological time on each island (need to include island ages somewhere)
}
