% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{duplicate}
\alias{duplicate}
\title{Duplicate an object.  This has an added effect of turning an ALTREP object into a regular object}
\usage{
duplicate(x)
}
\arguments{
\item{x}{object}
}
\value{
duplicate of the given object
}
\description{
Duplicate an object.  This has an added effect of turning an ALTREP object into a regular object
}
