% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-fma.R
\name{fmadd}
\alias{fmadd}
\alias{fmsub}
\alias{fnmadd}
\alias{fnmsub}
\title{In-situ Fused multiply add}
\usage{
fmadd(x, a, b)

fmsub(x, a, b)

fnmadd(x, a, b)

fnmsub(x, a, b)
}
\arguments{
\item{x, a, b}{vectors of the same length}
}
\value{
\code{x} modified in-place and retuned invisibly
}
\description{
\code{x = x * a + b}
}
\examples{
x <- c(1, 2, 3)
a <- c(2, 3, 4)
b <- c(10, 10, 10)
fmadd(x, a, b)
x
}
