% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_outlier.R
\name{id_outlier}
\alias{id_outlier}
\title{Identify outliers via point and click on temp vs salinity plot}
\usage{
id_outlier(
  data,
  datecol = "date_time",
  tempcol = "temperature",
  salcol = "salinity",
  rowseq = seq(1:nrow(data)),
  plot_type = "sal/time",
 
    plot_title = "Manually identify anomalous data points by clicking on them. \\n Click stop on the upper left and exit for the function to return row numbers of ID'ed points"
)
}
\arguments{
\item{data}{(data.frame) Data frame with columns for date or temperature, salinity.}

\item{datecol}{(character) Name of colume for date time. Defaults to "date_time".}

\item{tempcol}{(character) Name of column for temperature. Defaults to "temperature".}

\item{salcol}{(character) Name of column for salinity. Defaults to "salinity".}

\item{rowseq}{(numeric) Numeric vector of row numbers starting from one. Defaults to seq(1:nrow(data)).}

\item{plot_type}{(character) Either "sal/time" for plotting salinity over time, or "temp/sal" for plotting temperature vs salinity. Defaults to "sal/time".}

\item{plot_title}{(character) Give a plot title. Defaults to "Manually identify anomalous data points by clicking on them. Click stop on the upper left and exit for the function to return row numbers of ID'ed points". This is useful only if you'd like to save the plot as image.}
}
\value{
Numeric vector of row numbers manually ID'ed by user. Note that this returns an error if user did not click stop before exiting the plot application.
}
\description{
Identify outliers via point and click on temp vs salinity plot
}
