% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_island_nodes.R
\name{identify_island_nodes}
\alias{identify_island_nodes}
\title{Flags internal nodes as within-island or between-island based on descendant tip composition (no ASR)}
\usage{
identify_island_nodes()
}
\description{
Flags internal nodes as within-island or between-island based on descendant tip composition (no ASR)
}
