% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSoutliers_functions.R
\name{IDoutlier2}
\alias{IDoutlier2}
\title{Identify outliers via point and click on temp vs salinity plot}
\usage{
IDoutlier2(dat, tempname, salname, dfname)
}
\arguments{
\item{dat}{dataframe with columns for date or temperature, salinity, and row (has to be called row)}

\item{tempname}{name of column for temperature, in quotes}

\item{salname}{name of column for salinity, in quotes}

\item{dfname}{plot title}
}
\description{
Identify outliers via point and click on temp vs salinity plot
}
