% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Import data from instrument-generated files.}
\usage{
import_data(file_list, data_type, instrument_type, param_type,
  column_names = c("Temperature", "Conductivity", "Salinity"))
}
\arguments{
\item{file_list}{(character) Character vector of file paths. Use list.files() with full.names set to TRUE for this.}

\item{data_type}{(character) Name of data type. Processing will differ according to data type. Accepted input: "CTD" for conductivity-temperature-depth, or "TCM" for tilt current meter.}

\item{instrument_type}{(character) Name of instrument (only applies for "CTD" data type). Processing will differ according to instrument. Accepted input: "RBR", "SO".}

\item{param_type}{(character) Name of parameter type (only applies for "TCM" data type). Processing will differ according to parameter. Accepted inputs: "temp", "current".}

\item{column_names}{(character) Optional. This actually doesn't do anything atm FYI.}
}
\description{
Import data from instrument-generated files.
}
