% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-ops-unary.R
\name{ins_abs}
\alias{ins_abs}
\alias{ins_sqrt}
\alias{ins_floor}
\alias{ins_ceil}
\alias{ins_trunc}
\alias{ins_round}
\alias{ins_exp}
\alias{ins_log}
\alias{ins_cos}
\alias{ins_sin}
\alias{ins_tan}
\alias{ins_not}
\alias{ins_expm1}
\alias{ins_log1p}
\alias{ins_acos}
\alias{ins_asin}
\alias{ins_atan}
\alias{ins_acosh}
\alias{ins_asinh}
\alias{ins_atanh}
\alias{ins_cosh}
\alias{ins_sinh}
\alias{ins_tanh}
\alias{ins_cospi}
\alias{ins_sinpi}
\alias{ins_tanpi}
\alias{ins_sign}
\alias{ins_cumsum}
\alias{ins_cumprod}
\alias{ins_cummax}
\alias{ins_cummin}
\title{Math ops}
\usage{
ins_abs(x)

ins_sqrt(x)

ins_floor(x)

ins_ceil(x)

ins_trunc(x)

ins_round(x)

ins_exp(x)

ins_log(x)

ins_cos(x)

ins_sin(x)

ins_tan(x)

ins_not(x)

ins_expm1(x)

ins_log1p(x)

ins_acos(x)

ins_asin(x)

ins_atan(x)

ins_acosh(x)

ins_asinh(x)

ins_atanh(x)

ins_cosh(x)

ins_sinh(x)

ins_tanh(x)

ins_cospi(x)

ins_sinpi(x)

ins_tanpi(x)

ins_sign(x)

ins_cumsum(x)

ins_cumprod(x)

ins_cummax(x)

ins_cummin(x)
}
\arguments{
\item{x}{vector}
}
\description{
Math ops
}
