% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-ops.R
\name{ins_add}
\alias{ins_add}
\alias{\%+=\%}
\alias{ins_sub}
\alias{\%-=\%}
\alias{ins_mul}
\alias{\%*=\%}
\alias{ins_div}
\alias{\%/=\%}
\title{In-situ ops}
\usage{
ins_add(x, y)

x \%+=\% y

ins_sub(x, y)

x \%-=\% y

ins_mul(x, y)

x \%*=\% y

ins_div(x, y)

x \%/=\% y
}
\arguments{
\item{x, y}{params}
}
\value{
\code{x} modified in-place and returned invisibly
}
\description{
In-situ ops
}
