% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-copy.R
\name{ins_copy}
\alias{ins_copy}
\title{Copy segment}
\usage{
ins_copy(x, y, n = NULL, xi = 1L, yi = 1L)
}
\arguments{
\item{x, y}{numeric vectors}

\item{n}{number of elements to copy}

\item{xi, yi}{starting indices}
}
\description{
Copy segment
}
