% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-copy.R
\name{ins_copy_from}
\alias{ins_copy_from}
\title{Copy segment}
\usage{
ins_copy_from(x, y, xi, yi, n)
}
\arguments{
\item{x, y}{numeric vectors}

\item{xi, yi}{starting indices}

\item{n}{number of elements to copy}
}
\description{
Copy segment
}
