% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-copy.R
\name{ins_copy_if}
\alias{ins_copy_if}
\title{Copy all or part of one vector into another}
\usage{
ins_copy_if(x, y, lgl)
}
\arguments{
\item{x}{destination vector.  Will be modified in-place}

\item{y}{source. Can be scalar of vector of values}

\item{lgl}{numeric vector where 0 represents false, and all non-zero values
represent true}
}
\value{
x modified in-place
}
\description{
Copy all or part of one vector into another
}
