% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-ops.R
\name{ins_plus}
\alias{ins_plus}
\alias{\%+=\%}
\alias{ins_minus}
\alias{\%-=\%}
\alias{ins_mult}
\alias{\%*=\%}
\alias{ins_divide}
\alias{\%/=\%}
\title{In-situ ops}
\usage{
ins_plus(x, y)

x \%+=\% y

ins_minus(x, y)

x \%-=\% y

ins_mult(x, y)

x \%*=\% y

ins_divide(x, y)

x \%/=\% y
}
\arguments{
\item{x, y}{params}
}
\value{
\code{x} modified in-place and returned invisibly
}
\description{
In-situ ops
}
