% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-runif.R
\name{ins_runif}
\alias{ins_runif}
\title{Write random values into a vector in-situ. Uses xoshiro256++ for RNG}
\usage{
ins_runif(x, lower, upper)
}
\arguments{
\item{x}{vector of real, integer values}

\item{lower, upper}{limits of random numbers generated}
}
\description{
Write random values into a vector in-situ. Uses xoshiro256++ for RNG
}
