% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-sort.R
\name{ins_sort}
\alias{ins_sort}
\title{Sort a vector of real values in-situ}
\usage{
ins_sort(x, decreasing = FALSE)
}
\arguments{
\item{x}{vector of real values}

\item{decreasing}{order. default FALSE}
}
\description{
Sort a vector of real values in-situ
}
