% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-package.R
\docType{package}
\name{insitu-package}
\alias{insitu}
\alias{insitu-package}
\title{insitu: Estimating in situ Speciation on Island Systems}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://datadiversitylab.github.io/insitu/}
}

}
\author{
\strong{Maintainer}: Kristen Martinet \email{kmartinet@arizona.edu}

Authors:
\itemize{
  \item Cristian Román-Palacios \email{cromanpa@arizona.edu} [contributor]
}

}
\keyword{internal}
