% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-zzz.R
\docType{data}
\name{insitu}
\alias{insitu}
\title{List of all insitu functions}
\format{
An object of class \code{list} of length 54.
}
\usage{
insitu
}
\description{
List of all insitu functions
}
\keyword{datasets}
