% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu_confidence.R
\name{insitu_confidence}
\alias{insitu_confidence}
\title{Computes per-node support for in-situ classification}
\usage{
insitu_confidence()
}
\description{
Computes per-node support for in-situ classification
}
