% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-fill.R
\name{insitu_fill}
\alias{insitu_fill}
\title{Write value into a vector in-situ}
\usage{
insitu_fill(x, value)
}
\arguments{
\item{x}{vector of real, integer or character values}

\item{value}{value to fill with vector. must be same type as \code{x}}
}
\description{
Write value into a vector in-situ
}
