% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-fill.R
\name{insitu_fill_runif}
\alias{insitu_fill_runif}
\title{Write random values into a vector in-situ}
\usage{
insitu_fill_runif(x, lower, upper)
}
\arguments{
\item{x}{vector of real, integer values}

\item{lower, upper}{limits of random numbers generated}
}
\description{
Write random values into a vector in-situ
}
