% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-replace.R
\name{insitu_replace}
\alias{insitu_replace}
\title{Write values into a vector in-situ}
\usage{
insitu_replace(x, idx, values)
}
\arguments{
\item{x}{vector of real, integer or character values}

\item{idx}{index at which to start writing}

\item{values}{values to insert. must be same type as \code{x}}
}
\description{
Write values into a vector in-situ
}
