% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu_richness_index.R
\name{insitu_richness_index}
\alias{insitu_richness_index}
\title{Computes an index of in-situ contribution to diversity (standardized), comparable across islands of different sizes and ages}
\usage{
insitu_richness_index()
}
\description{
Computes an index of in-situ contribution to diversity (standardized), comparable across islands of different sizes and ages
}
