% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-shuffle.R
\name{insitu_shuffle}
\alias{insitu_shuffle}
\title{Shuffle a vector in place}
\usage{
insitu_shuffle(x)
}
\arguments{
\item{x}{vector of values}
}
\description{
Uses fisher-yates.
}
\details{
The \emph{fast} variant of this function uses a xoshiro256++ PRNG which is
much faster than sourcing randomness from R's random number generator.
}
