% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu_speciation_index.R
\name{insitu_speciation_index}
\alias{insitu_speciation_index}
\title{In-situ speciation index per island}
\usage{
insitu_speciation_index(phy, events)
}
\arguments{
\item{phy}{Matched phylogenetic tree}

\item{events}{The data frame returned by \code{map_insitu_events}}
}
\value{
A data frame with one row per island and columns:
\describe{
\item{\code{island}}{Island name.}
\item{\code{n_insitu}}{Number of in-situ speciation events.}
\item{\code{island_bl}}{Total branch length of the island subtree.}
\item{\code{isci}}{In-situ contribution index: \code{n_insitu / island_bl}.}
}
}
\description{
Computes a standardized index of in-situ contribution to diversity for
each island. The index (isci) is the number of in-situ speciation events
on an island divided by the total branch length of the island's subtree.
This is, the sum of all branch lengths of all descendant tips that are
exclusively assigned to that island.
}
