% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu_speciation_rate.R
\name{insitu_speciation_rate}
\alias{insitu_speciation_rate}
\title{Estimates per-lineage in-situ speciation rates from the phylogeny}
\usage{
insitu_speciation_rate()
}
\description{
Estimates per-lineage in-situ speciation rates from the phylogeny
}
