% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_insitu_events.R
\name{map_insitu_events}
\alias{map_insitu_events}
\title{Applies decision rules to reconstructed nodes to classify each speciation event as in-situ, colonization, or ambiguous}
\usage{
map_insitu_events(recons, phy, PAM, threshold = 0.5)
}
\arguments{
\item{recons}{A list of ancestral state reconstructions returned by \code{run_geo_asr}.
We do not recommend inputting a custom list for this function, but if you
need to, make sure that the list is composed of \code{ape::ace} objects, each
named after a locality. Refer to \code{run_geo_asr} for more information about
how the original function estimates ancestral states.}

\item{phy}{The phylogenetic tree associated with your data.}

\item{PAM}{A presence-absence matrix reflecting where each species of
interest is located. The \code{match_island_phylo} function will create this PAM,
but if the user would prefer to input a custom PAM, make sure that it has a
column titled "locale" with the name of an island in each row, and each
subsequent column is titled with a species name. Each species column should
include either a 0 (absence) or a 1 (presence), signifying whether that
species occurs on the island in a given row.}

\item{threshold}{The threshold of probability at which it is reasonable to
infer that a given ancestral node occurred on the same island as a given
set of tips. Default: 0.5.}
}
\value{
A data frame including node numbers, islands associated with those
nodes, whether an in-situ speciation event was likely at that node/island
combination, and for transitions, the source island (export) and
destination island (import).
}
\description{
Applies decision rules to reconstructed nodes to classify each speciation event as in-situ, colonization, or ambiguous
}
