% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_island_phylo.R
\name{match_island_phylo}
\alias{match_island_phylo}
\title{Imports and validates dated phylogenies with associated island occurrence data}
\usage{
match_island_phylo(phy, locs, exclude = TRUE)
}
\arguments{
\item{phy}{The phylogenetic tree associated with your data}

\item{locs}{A dataframe with two columns: "species" and "locale". Each row
should represent a single locality where that species is located (e.x., if
a species is found in multiple localities, there should be multiple rows for
that species). This package focuses on island occurrences, so if a species is
found on the mainland, write its locality as "Mainland".}

\item{exclude}{(Logical) Set to TRUE if you want to exclude mainland species
from the output. Default: TRUE}
}
\value{
A list with 3 items:
\itemize{
\item \code{PAM}: A presence-absence matrix describing where species occur
\item \code{sp_df}: A dataframe summarizing what species, and how many, are on each island
\item \code{phy}: A pruned version of the user-provided phylogeny that includes only species that are in the user-provided locality data
}
}
\description{
Imports and validates dated phylogenies with associated island occurrence data
}
\examples{
# Read tree trimmed from Patton et al. 2021
tree <- ape::read.tree(system.file("extdata",
                                   "Patton_etal_trimmed.tree",
                                    package = "insitu"))
# Read species location dataframe
dat <- read.csv(system.file("extdata",
                            "anolis_dat.csv",
                             package = "insitu"))
# Match island data and phylogeny
matched <- match_island_phylo(phy = tree, locs = dat)

}
