% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSoutliers_functions.R
\name{outlierTS}
\alias{outlierTS}
\title{Visulaize the outliers (black) in plot over time}
\usage{
outlierTS(dat, tempcol, name)
}
\arguments{
\item{dat}{dataframe that has gone through TSanom_workup}

\item{tempcol}{name of column for temperature, in quotes}

\item{name}{name of plot, usually site name}
}
\description{
Visulaize the outliers (black) in plot over time
}
