% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_classified_phylo.R
\name{plot_classified_phylo}
\alias{plot_classified_phylo}
\title{Dated phylogeny with branches and nodes colored by in-situ vs. colonization classification}
\usage{
plot_classified_phylo()
}
\description{
Dated phylogeny with branches and nodes colored by in-situ vs. colonization classification
}
