% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diversity_decomposition.R
\name{plot_diversity_decomposition}
\alias{plot_diversity_decomposition}
\title{Stacked bar or area chart of in-situ vs. colonization diversity per island (maybe through time?)}
\usage{
plot_diversity_decomposition()
}
\description{
Stacked bar or area chart of in-situ vs. colonization diversity per island (maybe through time?)
}
