% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_island_map.R
\name{plot_island_map}
\alias{plot_island_map}
\title{Map with per-island diversity decomposition metrics displayed spatially}
\usage{
plot_island_map()
}
\description{
Map with per-island diversity decomposition metrics displayed spatially
}
