% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_island_nodes.R
\name{plot_island_nodes}
\alias{plot_island_nodes}
\title{Plot island node classification on a phylogeny}
\usage{
plot_island_nodes(phy, node_class, PAM)
}
\arguments{
\item{phy}{The matched phylogenetic tree .}

\item{node_class}{The data frame returned by \code{identify_island_nodes}.}

\item{PAM}{A presence-absence matrix.}
}
\description{
Plot island node classification on a phylogeny
}
