% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TS.R
\name{plot_tempsal}
\alias{plot_tempsal}
\title{Initial look at Temperature and Salinity data}
\usage{
plot_tempsal(df, Temp, Sal, Date, Pres = 1.3, plottitle = "T vs S")
}
\arguments{
\item{Temp}{(character) column name of temperature values in Celsius}

\item{Sal}{(character) column name of salinity}

\item{Date}{(dttm) column name of date or date time}

\item{plottitle}{(character) title of plot, default is "T vs S"}

\item{P}{(numeric) pressure in atm, default is 1}
}
\value{
side by side plots, first of temperature vs salinity with freezing line, second of salinity over time, third of temperature over time
}
\description{
Returns figure with plots of temperature vs salinity,  salinity over time, and temperature over time
}
