% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_island_phylo.R
\name{read_island_phylo}
\alias{read_island_phylo}
\title{Imports and validates dated phylogenies with associated island occurrence data}
\usage{
read_island_phylo(phy_path, occ_path)
}
\arguments{
\item{phy_path}{The local file path to the tree file}

\item{occ_path}{The local file path to the occurrence data file}
}
\value{
A list with 2 items: the pruned phylogenetic tree and associated data
}
\description{
Requirements: make sure that the row names for your data are the species names
}
