% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random64.R
\name{set_seed_lehmer}
\alias{set_seed_lehmer}
\title{Initialise the state of this package's Lehmer RNG}
\usage{
set_seed_lehmer(seed)
}
\arguments{
\item{seed}{initeger seed value}
}
\value{
None.
}
\description{
Initialise the state of this package's Lehmer RNG
}
\examples{
set_seed_lehmer(1)
x <- alloc_n(10)
br_runif(x)
x
}
