% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random64.R
\name{set_seed_random64}
\alias{set_seed_random64}
\title{Initialise the state of the internal RNG based upon R's current seed}
\usage{
set_seed_random64()
}
\description{
Initialise the state of the internal RNG based upon R's current seed
}
