% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-2d.R
\name{tf2_add_rotate}
\alias{tf2_add_rotate}
\title{Add rotation to a transformation matrix}
\usage{
tf2_add_rotate(mat, theta)
}
\arguments{
\item{mat}{3x3 transformation matrix}

\item{theta}{rotation angle (radians)}
}
\value{
None. \code{mat} modified by reference and returned invisibly
}
\description{
Add rotation to a transformation matrix
}
\examples{
tf <- tf2_new()
tf2_add_rotate(tf, pi/6)
tf
}
