% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-2d.R
\name{tf2_add_scale}
\alias{tf2_add_scale}
\title{Add scaling to a transformation matrix}
\usage{
tf2_add_scale(mat, x = 1, y = x)
}
\arguments{
\item{mat}{3x3 transformation matrix}

\item{x, y}{scaling}
}
\value{
None. \code{mat} modified by reference and returned invisibly
}
\description{
Add scaling to a transformation matrix
}
\examples{
tf <- tf2_new()
tf2_add_scale(tf, 1, 2)
tf
}
