% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-2d.R
\name{tf2_add_translate}
\alias{tf2_add_translate}
\title{Add translation to a transformation matrix}
\usage{
tf2_add_translate(mat, x = 0, y = 0)
}
\arguments{
\item{mat}{3x3 transformation matrix}

\item{x, y}{translation}
}
\value{
None. \code{mat} modified by reference and returned invisibly
}
\description{
Add translation to a transformation matrix
}
\examples{
tf <- tf2_create()
tf2_add_translate(tf, 1, 2)
tf
}
