% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-2d.R
\name{tf2_create}
\alias{tf2_create}
\title{Create identity transform for 2-D}
\usage{
tf2_create()
}
\value{
3x3 identity matrix
}
\description{
Create identity transform for 2-D
}
\examples{
tf2_create()
}
