% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-2d.R
\name{tf2_reset}
\alias{tf2_reset}
\title{Reset a 2-D transformation matrix back to the identity matrix}
\usage{
tf2_reset(mat)
}
\arguments{
\item{mat}{3x3 transformation matrix}
}
\value{
None. \code{mat} modified by reference and returned invisibly
}
\description{
Reset a 2-D transformation matrix back to the identity matrix
}
\examples{
tf <- matrix(1, 3, 3)
tf2_reset(tf)
tf
}
