% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-3d.R
\name{tf3_add_scale}
\alias{tf3_add_scale}
\title{Add scaling to a transformation matrix}
\usage{
tf3_add_scale(mat, x = 1, y = x, z = x)
}
\arguments{
\item{mat}{4x4 transformation matrix}

\item{x, y, z}{scaling}
}
\value{
None. \code{mat} modified by reference and returned invisibly
}
\description{
Add scaling to a transformation matrix
}
\examples{
tf <- tf3_create()
tf3_add_scale(tf, 1, 2, 3)
tf
}
