% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-3d.R
\name{tf3_apply}
\alias{tf3_apply}
\title{Apply a 3D transform to a matrix of coordinates}
\usage{
tf3_apply(mat, tf)
}
\arguments{
\item{mat}{matrix with at least 4 columns representing (x, y, z) and d}

\item{tf}{4x4 transformation matrix}
}
\value{
None. \code{mat} is modified by reference and returned invisibly
}
\description{
Apply a 3D transform to a matrix of coordinates
}
\examples{
tf <- tf3_create()
tf <- tf3_add_translate(tf, 1, 2, 3)
mat <- cbind(
  x = as.numeric(1:6),
  y = as.numeric(1:6),
  z = as.numeric(1:6),
  d = 1
)
tf
mat

tf3_apply(mat, tf)
mat
}
