% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-3d.R
\name{tf3_reset}
\alias{tf3_reset}
\title{Reset a transformation matrix back to the identity matrix}
\usage{
tf3_reset(mat)
}
\arguments{
\item{mat}{4x4 transformation matrix}
}
\value{
None. \code{mat} modified by reference and returned invisibly
}
\description{
Reset a transformation matrix back to the identity matrix
}
\examples{
tf <- matrix(1, 4, 4)
tf3_reset(tf)
tf
}
