% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tf_add_rotate_z}
\alias{tf_add_rotate_z}
\title{Add a scale transformation to a given matrix}
\usage{
tf_add_rotate_z(mat, theta)
}
\arguments{
\item{mat}{4x4 transformation matrix}

\item{theta}{rotation angle (radians)}
}
\value{
None. \code{mat} modified by reference and returned invisibly
}
\description{
Add a scale transformation to a given matrix
}
\examples{
mat <- tf_create()
tf_add_rotate_z(mat, pi/6)
mat
}
