% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tf_add_scale}
\alias{tf_add_scale}
\title{Add a scale transformation to a given matrix}
\usage{
tf_add_scale(mat, x = 1, y = x, z = x)
}
\arguments{
\item{mat}{4x4 transformation matrix}

\item{x, y, z}{translation}
}
\value{
None. \code{mat} modified by reference and returned invisibly
}
\description{
Add a scale transformation to a given matrix
}
\examples{
mat <- tf_create_identity()
tf_add_scale(mat, 1, 2, 3)
mat
}
