% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tf_translate}
\alias{tf_translate}
\title{Add a translation transformation to a given matrix}
\usage{
tf_translate(mat, x = 0, y = 0, z = 0)
}
\arguments{
\item{mat}{4x4 transformation matrix}

\item{x, y, z}{translation}
}
\value{
None. \code{mat} modified by reference and returned invisibly
}
\description{
Add a translation transformation to a given matrix
}
\examples{
mat <- identity(4)
tf_translate(mat, 1, 2 3)
mat
}
