% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSoutliers_functions.R
\name{TSanom_workup}
\alias{TSanom_workup}
\title{Identify invalid ("anom") data based on freezing line
calculates the salinity error based on the innate loggger cond and temp error margins (Terror and Cerror)
determines if dataset's error margins fall within an acceptable range
data is "anomylous" if data +C+T error is below the freezing line}
\usage{
TSanom_workup(dat, tempcol, condcol, Terror, Cerror)
}
\arguments{
\item{dat}{dataframe with columns for temperature and conductivity}

\item{tempcol}{name of column for temperature, in quotes}

\item{condcol}{name of column for conductivity, in quotes}

\item{Terror}{precision for temperature}

\item{Cerror}{precision for conductivity}
}
\description{
Identify invalid ("anom") data based on freezing line
calculates the salinity error based on the innate loggger cond and temp error margins (Terror and Cerror)
determines if dataset's error margins fall within an acceptable range
data is "anomylous" if data +C+T error is below the freezing line
}
