% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TScal.R
\name{TScal}
\alias{TScal}
\title{Calibrate temperature and salinity data}
\usage{
TScal(instrument_data, ysi_data, station_colname = "station",
  calibrate_by = "value", raw = "temperature",
  calibrated = paste0(raw, "_calibrated"), some_missing = TRUE)
}
\arguments{
\item{instrument_data}{(data.frame) Data frame of instrument data (that needs to be calibrated). Required column names are "date_time", "temperature", "conductivity". "date_time" column needs to be in POSIXct format.}

\item{ysi_data}{(data.frame) Data frame of measurements to calibrate by (often by a YSI). Required column names are "date_time", "temperature", "conductivity". "date_time" column needs to be in POSIXct format.}

\item{station_colname}{(character) Name of station code column in YSI data. Defaults to "station".}

\item{calibrate_by}{(character) Name of value column in YSI data (data to calibrate by). Defaults to "value". This column and the raw column should be measurements of the same thing, e.g. temperature.}

\item{raw}{(character) Name of value column in instrument data (raw data). Defaults to "temperature".}

\item{calibrated}{(character) Name of column to store calibrated data. Default to "_calibrated" appended to the raw column.}

\item{some_missing}{(logical) T/F on whether to calibrate where there are beginning or end dates that do not have calibrated values (e.g. if the instrument was put down on June 1 but there's no calibration data until July 1). If TRUE, function will assume the instrument data on the first/last dates are true and use them as calibration points. If FALSE, the time portions prior or after the first/last calibration values will not altered. Defaults to TRUE.}
}
\description{
Calibrate temperature and salinity data
}
