% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TScalibration_functions.R
\name{TScal_RBR}
\alias{TScal_RBR}
\title{Calibrate RBR data
Based on initial calibration measurment and end calibration measurement}
\usage{
TScal_RBR(dat)
}
\arguments{
\item{dat}{is a dataframe with rows:Conductivity, Temperature, fixedCond, fixedTemp}
}
\value{
calibrated dataframe
}
\description{
Calibrate RBR data
Based on initial calibration measurment and end calibration measurement
}
