% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TScalibration_functions.R
\name{TScal_SO}
\alias{TScal_SO}
\title{Calibrate StarOddi data 
Based on initial calibration measurment and end calibration measurement}
\usage{
TScal_SO(dat)
}
\arguments{
\item{dat}{a dataframe with rows:Cond_mspercm, Temp_C, fixedCond, fixedTemp}
}
\value{
calibrated dataframe
}
\description{
Calibrate StarOddi data 
Based on initial calibration measurment and end calibration measurement
}
