% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TScalibration_functions.R
\name{TSmidcal_RBR}
\alias{TSmidcal_RBR}
\title{Calibrate RBR data with midpoint
Based on initial calibration measurment, a midpoint measurement, and end calibration measurement}
\usage{
TSmidcal_RBR(dat, midpoint)
}
\arguments{
\item{dat}{is a dataframe with rows:Conductivity, Temperature, fixedCond, fixedTemp}

\item{midpoint}{the row number of the midpoint calibration}
}
\value{
calibrated dataframe
}
\description{
Calibrate RBR data with midpoint
Based on initial calibration measurment, a midpoint measurement, and end calibration measurement
}
