% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TScalibration_functions.R
\name{TSmidcal_SO}
\alias{TSmidcal_SO}
\title{Calibrate StarOddi data with midpoint
Based on initial calibration measurment, a midpoint measurement, and end calibration measurement}
\usage{
TSmidcal_SO(dat, midpoint)
}
\arguments{
\item{dat}{a dataframe with rows:Cond_mspercm, Temp_C, fixedCond, fixedTemp}

\item{midpoint}{the row number of the midpoint calibration}
}
\value{
calibrated dataframe
}
\description{
Calibrate StarOddi data with midpoint
Based on initial calibration measurment, a midpoint measurement, and end calibration measurement
}
