% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-ops-unary-y.R
\name{ybr_sqrt}
\alias{ybr_sqrt}
\title{Math operations taking only a single input vector}
\usage{
ybr_sqrt(x, idx = NULL, where = NULL, cols = NULL)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}

\item{idx}{vector of indices}

\item{where}{logical vector stored as floating point values. 0 = FALSE, 
all non-zero values treated as TRUE. Default: NULL.
This value indicates if the operation should be performed for the 
corresponding element in \code{x}.}

\item{cols}{which matrix columns should this apply to?  Argument only valid
if \code{x} is a matrix}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
Math operations taking only a single input vector
}
\examples{
# By-reference \code{abs()}
x <- c(-1, 0, 1)
br_abs(x)
x

# Conditional application of sqrt()
x <- as.numeric(-5:5)
t <- duplicate(x)     # will use this to hold a numeric logical vector
br_gt(t, 0)           # where is the value > 0
br_sqrt(x, where = t) # only sqrt(x) where x > 0
x
}
