% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc.R
\name{alloc_n}
\alias{alloc_n}
\alias{alloc_for}
\title{Allocate a new numeric vector}
\usage{
alloc_n(n)

alloc_for(x)
}
\arguments{
\item{n}{length of new vector}

\item{x}{new vector will be the same length as this vector}
}
\value{
new vector of the required length
}
\description{
These functions are like \code{numeric()} except the contents of the vector
are not initialized
}
\examples{
alloc_n(10)
}
