% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-ops-unary.R
\name{br_abs}
\alias{br_abs}
\alias{br_sqrt}
\alias{br_floor}
\alias{br_ceil}
\alias{br_trunc}
\alias{br_round}
\alias{br_exp}
\alias{br_log}
\alias{br_cos}
\alias{br_sin}
\alias{br_tan}
\alias{br_not}
\alias{br_expm1}
\alias{br_log1p}
\alias{br_acos}
\alias{br_asin}
\alias{br_atan}
\alias{br_acosh}
\alias{br_asinh}
\alias{br_atanh}
\alias{br_cosh}
\alias{br_sinh}
\alias{br_tanh}
\alias{br_cospi}
\alias{br_sinpi}
\alias{br_tanpi}
\alias{br_sign}
\alias{br_cumsum}
\alias{br_cumprod}
\alias{br_cummax}
\alias{br_cummin}
\alias{br_log2}
\alias{br_log10}
\alias{br_is_na}
\title{Math operations taking only a single input vector}
\usage{
br_abs(x, where = NULL)

br_sqrt(x, where = NULL)

br_floor(x, where = NULL)

br_ceil(x, where = NULL)

br_trunc(x, where = NULL)

br_round(x, digits)

br_exp(x, where = NULL)

br_log(x, where = NULL)

br_cos(x, where = NULL)

br_sin(x, where = NULL)

br_tan(x, where = NULL)

br_not(x, where = NULL)

br_expm1(x, where = NULL)

br_log1p(x, where = NULL)

br_acos(x, where = NULL)

br_asin(x, where = NULL)

br_atan(x, where = NULL)

br_acosh(x, where = NULL)

br_asinh(x, where = NULL)

br_atanh(x, where = NULL)

br_cosh(x, where = NULL)

br_sinh(x, where = NULL)

br_tanh(x, where = NULL)

br_cospi(x, where = NULL)

br_sinpi(x, where = NULL)

br_tanpi(x, where = NULL)

br_sign(x, where = NULL)

br_cumsum(x, where = NULL)

br_cumprod(x, where = NULL)

br_cummax(x, where = NULL)

br_cummin(x, where = NULL)

br_log2(x, where = NULL)

br_log10(x, where = NULL)

br_is_na(x, where = NULL)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}

\item{where}{logical vector stored as floating point values. 0 = FALSE, 
all non-zero values treated as TRUE. Default: NULL.
This value indicates if the operation should be performed for the 
corresponding element in \code{x}.}

\item{digits}{number of decimal places}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
Math operations taking only a single input vector
}
\examples{
# By-reference \code{abs()}
x <- c(-1, 0, 1)
br_abs(x)
x

# Conditional application of sqrt()
x <- as.numeric(-5:5)
t <- duplicate(x)     # will use this to hold a numeric logical vector
br_gt(t, 0)           # where is the value > 0
br_sqrt(x, where = t) # only sqrt(x) where x > 0
x
}
