% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-ops-binary.R
\name{br_add}
\alias{br_add}
\alias{br_sub}
\alias{br_mul}
\alias{br_div}
\alias{br_pow}
\alias{br_eq}
\alias{br_ne}
\alias{br_lt}
\alias{br_le}
\alias{br_gt}
\alias{br_ge}
\alias{br_and}
\alias{br_or}
\alias{br_rem}
\alias{br_idiv}
\alias{br_max}
\alias{br_min}
\alias{br_assign}
\alias{br_distsq}
\title{Math operations taking two arguments}
\usage{
br_add(x, y, idx = NULL, where = NULL, cols = NULL)

br_sub(x, y, idx = NULL, where = NULL, cols = NULL)

br_mul(x, y, idx = NULL, where = NULL, cols = NULL)

br_div(x, y, idx = NULL, where = NULL, cols = NULL)

br_pow(x, y, idx = NULL, where = NULL, cols = NULL)

br_eq(x, y, idx = NULL, where = NULL, cols = NULL)

br_ne(x, y, idx = NULL, where = NULL, cols = NULL)

br_lt(x, y, idx = NULL, where = NULL, cols = NULL)

br_le(x, y, idx = NULL, where = NULL, cols = NULL)

br_gt(x, y, idx = NULL, where = NULL, cols = NULL)

br_ge(x, y, idx = NULL, where = NULL, cols = NULL)

br_and(x, y, idx = NULL, where = NULL, cols = NULL)

br_or(x, y, idx = NULL, where = NULL, cols = NULL)

br_rem(x, y, idx = NULL, where = NULL, cols = NULL)

br_idiv(x, y, idx = NULL, where = NULL, cols = NULL)

br_max(x, y, idx = NULL, where = NULL, cols = NULL)

br_min(x, y, idx = NULL, where = NULL, cols = NULL)

br_assign(x, y, idx = NULL, where = NULL, cols = NULL)

br_distsq(x, y, idx = NULL, where = NULL, cols = NULL)
}
\arguments{
\item{x}{numeric vector. This vector will be modified by-reference to 
contain the result of the calculation}

\item{y}{Either scalar numeric value or numeric vector of the same 
length as \code{x}.}

\item{idx}{idx}

\item{where}{logical vector stored as floating point values. 0 = FALSE, 
all non-zero values treated as TRUE. Default: NULL.
This value indicates if the operation should be performed for the 
corresponding element in \code{x}.}

\item{cols}{columns}
}
\value{
\code{x} argument is modified by-reference and returned invisibly
}
\description{
Math operations taking two arguments
}
\examples{
# x <- x + y
x <- as.numeric(1:10)
y <- as.numeric(1:10)
br_add(x, y)
x
# x <- x + 1
br_add(x, 1)
x
}
