% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-matrix.R
\name{br_mat_vec_mul_asq}
\alias{br_mat_vec_mul_asq}
\title{Matrix-vector multiplication when A is square}
\usage{
br_mat_vec_mul_asq(A, x, alpha = 1, ta = FALSE)
}
\arguments{
\item{A}{Matrix}

\item{x}{vector}

\item{alpha}{Scaling factor for \code{A * x}. Default: 1.0}

\item{ta}{Should matrix 'A' be transposed? Default: FALSE}
}
\value{
None. x is modified by-reference and returned invisibly
}
\description{
This function exposes the general matrix-vector multiplication operation from R's
included BLAS.  \code{x = alpha * A * x}
}
\examples{
A <- matrix(1, 3, 3)
x <- rep(1, 3)

# Compare to R's method
A \%*\% x

# Calculate. By-reference. Overwriting 'x'
br_mat_vec_mul_asq(A, x)
x
}
