% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colonization_rate.R
\name{colonization_rate}
\alias{colonization_rate}
\title{Estimates the rate of successful colonization events per lineage per unit time}
\usage{
colonization_rate()
}
\description{
Estimates the rate of successful colonization events per lineage per unit time
}
