% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_diversity.R
\name{decompose_diversity}
\alias{decompose_diversity}
\title{Partitions island species richness into colonization-derived vs. in-situ-derived richness per island}
\usage{
decompose_diversity()
}
\description{
Partitions island species richness into colonization-derived vs. in-situ-derived richness per island
}
