% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/altrep.R
\name{duplicate}
\alias{duplicate}
\title{Duplicate an R object}
\usage{
duplicate(x)
}
\arguments{
\item{x}{Any R object}
}
\value{
duplicate of the given object
}
\description{
Duplicate an R object
}
\examples{
x <- c(1, 2, 3)
y <- duplicate(x)
y
}
