% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_salinity.R
\name{flag_salinity}
\alias{flag_salinity}
\title{Identify and flag invalid ("anomalous") data based on freezing line}
\usage{
flag_salinity(data, tempcol = "temperature", condcol = "conductivity",
  Terror, Cerror, flag_scheme = NULL)
}
\arguments{
\item{data}{(data.frame) Data to be flagged.}

\item{tempcol}{(character) Name of column for temperature. Defaults to "temperature".}

\item{condcol}{(character) Name of column for conductivity. Defaults to "conductivity".}

\item{Terror}{(numeric) Precision for temperature.}

\item{Cerror}{(numeric) Precision for conductivity.}

\item{flag_scheme}{(character) Scheme for flagging anomalous salinity values. Defaults to NULL and outputs a logical TRUE/FALSE column. Alternatively supply a vector of two character flags to denote good/bad data respectively, e.g. c("valid", "invalid").}
}
\value{
(data.frame) The input dataframe with new columns added: "anomalous" (logical) indicating anomalous values, "salinity" (named according to input argument salcol) for calculated salinity.
}
\description{
Calculates the salinity error based on the innate loggger conductivity and temperature error margins.
Determines if dataset's error margins fall within an acceptable range.
Data is "anomylous" if data +C+T error is below the freezing line.
}
