% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_salinity.R
\name{flag_salinity}
\alias{flag_salinity}
\title{Identify and flag invalid ("anomalous") data based on freezing line}
\usage{
flag_salinity(
  data,
  tempcol = "temperature",
  condcol = "conductivity",
  pressurecol,
  Terror,
  Cerror,
  ref_cond = 42.9,
  pressure_unit = "dbar",
  flag_colname = "anomalous",
  flag_scheme = c("valid", "invalid")
)
}
\arguments{
\item{data}{(data.frame) Data to be flagged.}

\item{tempcol}{(character) Name of column for temperature. Defaults to "temperature".}

\item{condcol}{(character) Name of column for conductivity. Defaults to "conductivity".}

\item{pressurecol}{(character) Name of column for pressure or depth. No default value.}

\item{Terror}{(numeric) Precision for temperature.}

\item{Cerror}{(numeric) Precision for conductivity.}

\item{ref_cond}{(numeric) Reference conductivity value for calculating salinity, defaults to 42.9 for conductivity measured in miliSeimens/m^2/s.}

\item{pressure_unit}{(character) Unit for pressure or depth for necessary conversion. Takes "bar" or "dbar". No other values accepted. Defaults to dbar.}

\item{flag_colname}{(character) Column name to give the new flag column, defaults to "anomalous".}

\item{flag_scheme}{(character) Scheme for flagging anomalous salinity values. Supply a vector of character flags to denote good/bad data respectively, e.g. c("valid", "invalid"). The first two will be used. IF NULL function will output a logical TRUE/FALSE column.}
}
\value{
(data.frame) The input dataframe with a new column added, named by the flag_colname argument, defaults to "anomalous" (character or logical) indicating anomalous values.
}
\description{
Calculates the salinity error based on the innate loggger conductivity and temperature error margins.
Determines if dataset's error margins fall within an acceptable range.
Data is "anomylous" if data +C+T error is below the freezing line.
}
