% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSoutliers_functions.R
\name{IDoutlier}
\alias{IDoutlier}
\title{Identify outliers via point and click on salinity over time plot}
\usage{
IDoutlier(dat, datename, salname, dfname)
}
\arguments{
\item{dat}{dataframe with columns for date or datetime, salinity, and row (has to be called row)}

\item{datename}{name of column for date, in quotes}

\item{salname}{name of column for salinity, in quotes}

\item{dfname}{plot title}
}
\description{
Identify outliers via point and click on salinity over time plot
}
