% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-math.R
\name{ins_abs}
\alias{ins_abs}
\alias{ins_sqrt}
\alias{ins_floor}
\alias{ins_ceil}
\alias{ins_trunc}
\alias{ins_round}
\alias{ins_exp}
\alias{ins_log}
\alias{ins_cos}
\alias{ins_sin}
\alias{ins_tan}
\title{Math ops}
\usage{
ins_abs(x)

ins_sqrt(x)

ins_floor(x)

ins_ceil(x)

ins_trunc(x)

ins_round(x)

ins_exp(x)

ins_log(x)

ins_cos(x)

ins_sin(x)

ins_tan(x)
}
\arguments{
\item{x}{vector}
}
\description{
Math ops
}
