% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-ops-binary.R
\name{ins_add}
\alias{ins_add}
\alias{\%+=\%}
\alias{ins_sub}
\alias{\%-=\%}
\alias{ins_mul}
\alias{\%*=\%}
\alias{ins_div}
\alias{\%/=\%}
\alias{ins_eq}
\alias{\%==\%}
\alias{ins_ne}
\alias{\%!=\%}
\alias{ins_lt}
\alias{\%<\%}
\alias{ins_le}
\alias{\%<=\%}
\alias{ins_gt}
\alias{\%>\%}
\alias{ins_ge}
\alias{\%>=\%}
\alias{ins_and}
\alias{\%&&\%}
\alias{ins_or}
\alias{\%||\%}
\alias{ins_rem}
\alias{ins_idiv}
\title{In-situ ops}
\usage{
ins_add(x, y)

x \%+=\% y

ins_sub(x, y)

x \%-=\% y

ins_mul(x, y)

x \%*=\% y

ins_div(x, y)

x \%/=\% y

ins_eq(x, y)

x \%==\% y

ins_ne(x, y)

x \%!=\% y

ins_lt(x, y)

x \%<\% y

ins_le(x, y)

x \%<=\% y

ins_gt(x, y)

x \%>\% y

ins_ge(x, y)

x \%>=\% y

ins_and(x, y)

x \%&&\% y

ins_or(x, y)

x \%||\% y

ins_rem(x, y)

ins_idiv(x, y)
}
\arguments{
\item{x, y}{params}
}
\value{
\code{x} modified in-place and returned invisibly
}
\description{
In-situ ops
}
