% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insitu-package.R
\docType{package}
\name{insitu-package}
\alias{insitu}
\alias{insitu-package}
\title{insitu: Estimating in situ Speciation on Island Systems}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Kristen Martinet \email{kmartinet@arizona.edu}

}
\keyword{internal}
