% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_insitu_events.R
\name{map_insitu_events}
\alias{map_insitu_events}
\title{Applies decision rules to reconstructed nodes to classify each speciation event as in-situ, colonization, or ambiguous}
\usage{
map_insitu_events()
}
\description{
Applies decision rules to reconstructed nodes to classify each speciation event as in-situ, colonization, or ambiguous
}
