% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_island_phylo.R
\name{read_island_phylo}
\alias{read_island_phylo}
\title{Imports and validates dated phylogenies with associated island occurence data}
\usage{
read_island_phylo()
}
\description{
Imports and validates dated phylogenies with associated island occurence data
}
