% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_geo_asr.R
\name{run_geo_asr}
\alias{run_geo_asr}
\title{Runs geographic state ASR (discrete: island; mainland?)}
\usage{
run_geo_asr(phy, PAM)
}
\arguments{
\item{phy}{The phylogenetic tree associated with your data}

\item{PAM}{A presence-absence matrix reflecting where each species of
interest is located. The \code{match_island_phylo} function will create this PAM,
but if the user would prefer to input a custom PAM, make sure that it has a
column titled "locale" with the name of an island in each row, and each
subsequent column is titled with a species name. Each species column should
include either a 0 (absence) or a 1 (presence), signifying whether that
species occurs on the island in a given row.}
}
\value{
A list of island-specific ancestral state reconstructions
}
\description{
Runs geographic state ASR (discrete: island; mainland?)
}
