% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmap_insitu.R
\name{simmap_insitu}
\alias{simmap_insitu}
\title{Uses stochastic character mapping to propagate ASR uncertainty through speciation classifications. Posterior distribution of in-situ event counts}
\usage{
simmap_insitu()
}
\description{
Uses stochastic character mapping to propagate ASR uncertainty through speciation classifications. Posterior distribution of in-situ event counts
}
