% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform-2d.R
\name{tf2_apply}
\alias{tf2_apply}
\title{Apply a 2-D transform to a matrix of coordinates}
\usage{
tf2_apply(mat, tf)
}
\arguments{
\item{mat}{matrix with at least 3 columns representing (x, y) and d}

\item{tf}{3x3 transformation matrix}
}
\value{
None. \code{mat} is modified by reference and returned invisibly
}
\description{
Apply a 2-D transform to a matrix of coordinates
}
\examples{
tf <- tf2_create()
tf <- tf2_add_translate(tf, 1, 2)
mat <- cbind(
  x = as.numeric(1:6),
  y = as.numeric(1:6),
  d = 1
)
tf
mat

tf2_apply(mat, tf)
mat
}
