% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{tf_add_rotate_x}
\alias{tf_add_rotate_x}
\alias{tf_add_rotate_y}
\alias{tf_add_rotate_z}
\title{Add rotation to a transformation matrix}
\usage{
tf_add_rotate_x(mat, theta)

tf_add_rotate_y(mat, theta)

tf_add_rotate_z(mat, theta)
}
\arguments{
\item{mat}{4x4 transformation matrix}

\item{theta}{rotation angle (radians)}
}
\value{
None. \code{mat} modified by reference and returned invisibly
}
\description{
Add rotation to a transformation matrix
}
\examples{
tf <- tf_create()
tf_add_rotate_x(tf, pi/6)
tf
}
