% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TScalibration_functions.R
\name{TS.adj}
\alias{TS.adj}
\title{Calibrate data
Based on initial calibration measurment and end calibration measurement
Adjustment is linear}
\usage{
TS.adj(raw, calib, rowfirst, rowlast, rownums)
}
\arguments{
\item{raw}{df column of raw data}

\item{calib}{df column that is blank except for the calibration values}

\item{rowfirst}{the rownumber of the first calibration value}

\item{rowlast}{the rownumber of the second (and last) calibration value}

\item{rownums}{a column of row numbers}
}
\value{
calibrated column calib
}
\description{
Calibrate data
Based on initial calibration measurment and end calibration measurement
Adjustment is linear
}
