% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS.adj.R, R/TScalibration_functions.R
\name{TS.adj}
\alias{TS.adj}
\title{Calibrate data
Based on initial calibration measurment and end calibration measurement. Adjustment is linear.}
\usage{
TS.adj(raw, calib, rowfirst, rowlast, rownums)

TS.adj(raw, calib, rowfirst, rowlast, rownums)
}
\arguments{
\item{raw}{(numeric) Numeric vector of raw data.}

\item{calib}{(numeric) df column that is blank except for the calibration values}

\item{rowfirst}{(numeric) the rownumber of the first calibration value}

\item{rowlast}{(numeric) the rownumber of the second (and last) calibration value}

\item{rownums}{a column of row numbers.}

\item{raw}{df column of raw data}

\item{calib}{df column that is blank except for the calibration values}

\item{rowfirst}{the rownumber of the first calibration value}

\item{rowlast}{the rownumber of the second (and last) calibration value}

\item{rownums}{a column of row numbers}
}
\value{
A vector with caliberated values. Adjustment is linear.

calibrated column calib
}
\description{
Calibrate data
Based on initial calibration measurment and end calibration measurement. Adjustment is linear.

Calibrate data
Based on initial calibration measurment and end calibration measurement
Adjustment is linear
}
